requirements.txt
fastapi
uvicorn
python-multipart
omnivoice
openai-whisper
aiogram
python-dotenv
requests
httpx
